requirements.txt
aiohttp
appdirs
async-timeout
attrs
black
certifi
cffi
chardet
charset-normalizer
click
colorama
cryptography
Flask
geoip2
idna
itsdangerous
Jinja2
jsonschema
MarkupSafe
maxminddb
multidict
mypy-extensions
passlib
pathspec
platformdirs
pycparser
PyJWT
pyotp
pyrsistent
qrcode
regex
requests
six
tomli
typing-extensions
urllib3
Werkzeug
yarl
pre-commit